Move some tests to run without multiple MPI ranks
