+++ unit_test/CMakeLists.txt
@@ -22,15 +22,17 @@ configure_file(${CMAKE_CURRENT_SOURCE_DIR}/../examples/Inp_SpotMelt.txt
 ## Create tests without Kokkos
 ##--------------------------------------------------------------------------##
 macro(ExaCA_add_tests_nobackend)
-  cmake_parse_arguments(EXACA_UNIT_TEST "" "" "NAMES" ${ARGN})
+  cmake_parse_arguments(EXACA_UNIT_TEST "MPI" "" "NAMES" ${ARGN})
   set(EXACA_UNIT_TEST_MPIEXEC_NUMPROCS 1)
-  foreach( _np 2 4 )
-    if(MPIEXEC_MAX_NUMPROCS GREATER_EQUAL ${_np})
-      list(APPEND EXACA_UNIT_TEST_MPIEXEC_NUMPROCS ${_np})
+  if(EXACA_UNIT_TEST_MPI)
+    foreach( _np 2 4 )
+      if(MPIEXEC_MAX_NUMPROCS GREATER_EQUAL ${_np})
+        list(APPEND EXACA_UNIT_TEST_MPIEXEC_NUMPROCS ${_np})
+      endif()
+    endforeach()
+    if(MPIEXEC_MAX_NUMPROCS GREATER 4)
+      list(APPEND EXACA_UNIT_TEST_MPIEXEC_NUMPROCS ${MPIEXEC_MAX_NUMPROCS})
     endif()
-  endforeach()
-  if(MPIEXEC_MAX_NUMPROCS GREATER 4)
-    list(APPEND EXACA_UNIT_TEST_MPIEXEC_NUMPROCS ${MPIEXEC_MAX_NUMPROCS})
   endif()
   set(EXACA_UNIT_TEST_MAIN unit_test_main_nobackend.cpp)
 
@@ -52,15 +54,17 @@ endmacro()
 ## Create main tests
 ##--------------------------------------------------------------------------##
 macro(ExaCA_add_tests)
-  cmake_parse_arguments(EXACA_UNIT_TEST "" "" "NAMES" ${ARGN})
+  cmake_parse_arguments(EXACA_UNIT_TEST "MPI" "" "NAMES" ${ARGN})
   set(EXACA_UNIT_TEST_MPIEXEC_NUMPROCS 1)
-  foreach( _np 2 4 )
-    if(MPIEXEC_MAX_NUMPROCS GREATER_EQUAL ${_np})
-      list(APPEND EXACA_UNIT_TEST_MPIEXEC_NUMPROCS ${_np})
+  if(EXACA_UNIT_TEST_MPI)
+    foreach( _np 2 4 )
+      if(MPIEXEC_MAX_NUMPROCS GREATER_EQUAL ${_np})
+        list(APPEND EXACA_UNIT_TEST_MPIEXEC_NUMPROCS ${_np})
+      endif()
+    endforeach()
+    if(MPIEXEC_MAX_NUMPROCS GREATER 4)
+      list(APPEND EXACA_UNIT_TEST_MPIEXEC_NUMPROCS ${MPIEXEC_MAX_NUMPROCS})
     endif()
-  endforeach()
-  if(MPIEXEC_MAX_NUMPROCS GREATER 4)
-    list(APPEND EXACA_UNIT_TEST_MPIEXEC_NUMPROCS ${MPIEXEC_MAX_NUMPROCS})
   endif()
   set(EXACA_UNIT_TEST_NUMTHREADS 1)
   foreach( _nt 2 4 )
@@ -103,4 +107,8 @@ endmacro()
 
 ExaCA_add_tests_nobackend(NAMES Init)
 
-ExaCA_add_tests(NAMES KokkosInit KokkosUpdate Print)
+ExaCA_add_tests_nobackend(MPI NAMES InitMPI)
+
+ExaCA_add_tests(NAMES Print)
+
+ExaCA_add_tests(MPI NAMES KokkosInit KokkosUpdate)
